Move updater into separate file
