+++ Vic2ToHoI4/Source/HOI4World/ScriptedTriggers/ScriptedTriggers.h
@@ -7,7 +7,6 @@
 #include "../../Configuration.h"
 #include "newParser.h"
 #include <map>
-#include <set>
 #include <vector>
 
 
@@ -23,17 +22,11 @@ class ScriptedTriggers: commonItems::parser
 		void replaceElectionsScriptedTrigger(std::string_view name, const std::string& replacementBody);
 		void replaceLawsWarSupportTriggers(const std::map<std::string_view, std::string>& replacements);
 
-		void updateScriptedTriggers(const std::set<std::string>& majorIdeologies);
-
 		[[nodiscard]] const auto& getIdeologyScriptedTriggers() const { return ideologyScriptedTriggers; }
 		[[nodiscard]] const auto& getElectionsScriptedTriggers() const { return electionsScriptedTriggers; }
 		[[nodiscard]] const auto& getLawsWarSupportTriggers() const { return lawsWarSupportTriggers; }
 	
 	private:
-		void updateIdeologyScriptedTriggers(const std::set<std::string>& majorIdeologies);
-		void updateElectionsScriptedTriggers(const std::set<std::string>& majorIdeologies);
-		void updateLawsWarSupportTriggers(const std::set<std::string>& majorIdeologies);
-	
 		std::vector<ScriptedTrigger> ideologyScriptedTriggers;
 		std::vector<ScriptedTrigger> electionsScriptedTriggers;
 		std::vector<ScriptedTrigger> lawsWarSupportTriggers;
