+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj
@@ -296,6 +296,7 @@
     <ClCompile Include="Source\HOI4World\Regiment.cpp" />
     <ClCompile Include="Source\HOI4World\ScriptedLocalisations\ScriptedLocalisations.cpp" />
     <ClCompile Include="Source\HOI4World\ScriptedTriggers\ScriptedTriggers.cpp" />
+    <ClCompile Include="Source\HOI4World\ScriptedTriggers\ScriptedTriggersUpdater.cpp" />
     <ClCompile Include="Source\HOI4World\ShipTypes\PossibleShipVariants.cpp" />
     <ClCompile Include="Source\HOI4World\ShipTypes\ShipModules.cpp" />
     <ClCompile Include="Source\HOI4World\ShipTypes\ShipVariant.cpp" />
@@ -471,6 +472,7 @@
     <ClInclude Include="Source\HOI4World\ScriptedLocalisations\ScriptedLocalisations.h" />
     <ClInclude Include="Source\HOI4World\ScriptedTriggers\ScriptedTrigger.h" />
     <ClInclude Include="Source\HOI4World\ScriptedTriggers\ScriptedTriggers.h" />
+    <ClInclude Include="Source\HOI4World\ScriptedTriggers\ScriptedTriggersUpdater.h" />
     <ClInclude Include="Source\HOI4World\ShipTypes\PossibleShipVariants.h" />
     <ClInclude Include="Source\HOI4World\ShipTypes\ShipModules.h" />
     <ClInclude Include="Source\HOI4World\ShipTypes\ShipVariant.h" />
